requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
atomicwrites==1.4.1
    # via dessert
attrs==23.1.0
    # via dessert
cabina==0.7.2
    # via vedro
dessert==1.4.7
    # via vedro
filelock==3.13.1
    # via vedro
greenlet==2.0.2
    # via playwright
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
munch==4.0.0
    # via dessert
niltype==0.3.4
    # via
    #   cabina
    #   vedro
pillow==10.1.0
    # via vedro-screenshot-matcher
pixelmatch==0.3.0
    # via vedro-screenshot-matcher
playwright==1.37.0
    # via
    #   -r requirements.in
    #   vedro-playwright
    #   vedro-screenshot-matcher
py==1.11.0
    # via dessert
pyee==9.0.4
    # via playwright
pygments==2.16.1
    # via rich
pyparsing==3.1.1
    # via vedro
pyqt5==5.15.10
    # via
    #   pyqtwebengine
    #   vedro-screenshot-matcher
pyqt5-qt5==5.15.2
    # via pyqt5
pyqt5-sip==12.13.0
    # via
    #   pyqt5
    #   pyqtwebengine
pyqtwebengine==5.15.6
    # via vedro-screenshot-matcher
pyqtwebengine-qt5==5.15.2
    # via pyqtwebengine
rich==13.6.0
    # via vedro
rtry==1.4.1
    # via vedro-screenshot-matcher
typing-extensions==4.8.0
    # via pyee
vedro==1.10.0
    # via
    #   -r requirements.in
    #   vedro-playwright
    #   vedro-screenshot-matcher
vedro-playwright==1.2.1
    # via -r requirements.in
vedro-screenshot-matcher==0.1.0
    # via -r requirements.in
